requirements.txt
python-telegram-bot>=22.0
anthropic>=0.34.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
